Task-3 - Split item actions to editModeAction and itemAction
