+++ onboarding-task/src/actions/itemActionCreators.js
@@ -1,4 +1,4 @@
-import { ADD_ITEM, DELETE_ITEM, UPDATE_ITEM, START_EDIT_ITEM, STOP_EDIT_ITEM } from './actionTypes.js';
+import { ADD_ITEM, DELETE_ITEM, UPDATE_ITEM } from './actionTypes.js';
 import { generateId } from '../utils/idGenerator.js';
 
 // do not export this one
@@ -35,23 +35,5 @@ function updateItem(id, text) {
   };
 }
 
-function startEditItem(id) {
-  return {
-    type: START_EDIT_ITEM,
-    payload: {
-      id,
-    },
-  };
-}
-
-function stopEditItem(id) {
-  return {
-    type: STOP_EDIT_ITEM,
-    payload: {
-      id,
-    },
-  };
-}
-
 const addItemWithDependencies = addItemFunctionCreator(generateId);
-export { addItemFunctionCreator, deleteItem, updateItem, startEditItem, addItemWithDependencies as addItem, stopEditItem };
+export { addItemFunctionCreator, deleteItem, updateItem, addItemWithDependencies as addItem };
